move globalvariables
