+++ portfolio-website/src/components/contact/ContactPageContainer.tsx
@@ -2,7 +2,7 @@ import { VStack } from '@chakra-ui/react'
 import { FC } from 'react'
 import { ParentProps } from '../props'
 import '@/css/animations.css'
-import { sizing } from '../shared/sizing'
+import { sizing } from '../../globalVariables'
 
 export const ContactPageContainer: FC<ParentProps> = ({ children }) => {
   return (
