+++ portfolio-website/src/globalVariables.ts
@@ -1,3 +1,5 @@
+export const svgColor = '#025e7b'
+
 export const sizing = {
   gap: {
     xlarge: { base: 10, md: 12, lg: 16 },
